+++ skelleton/sf/sql/_SHARE_CREATE.sql
@@ -1,7 +1,0 @@
-USE @@SF_DATABASEID@@;
-CREATE SHARE IF NOT EXISTS @@SF_SHARE_PUBLIC@@;
-grant usage on database @@SF_DATABASEID@@ to share @@SF_SHARE_PUBLIC@@;
-grant usage on schema @@SF_DATABASEID@@.@@SF_SCHEMA_SQUELLETON@@ to share @@SF_SHARE_PUBLIC@@;
-
-grant usage on function @@SF_DATABASEID@@.@@SF_SCHEMA_SQUELLETON@@.ST_BUFFER(GEOGRAPHY, DOUBLE, STRING, INT) to share @@SF_SHARE_PUBLIC@@;
-grant usage on function @@SF_DATABASEID@@.@@SF_SCHEMA_SQUELLETON@@.VERSION() to share @@SF_SHARE_PUBLIC@@;