+++ transform/library/js/transform_version.js
@@ -1,0 +1,11 @@
+// -----------------------------------------------------------------------
+// --
+// -- Copyright (C) 2021 CARTO
+// --
+// -----------------------------------------------------------------------
+
+/* exported transformVersion */
+transformVersion = function() {
+    return '1.0.0';
+}
+module.exports.transformVersion = transformVersion;